+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- **kiarina-utils-file**: Fixed symbolic link handling in read/write operations
+
 ## [1.2.0] - 2025-09-25
 
 ### Added